make a11y checks separate file
